Clone and remove a step from a step bundle

+++ source/javascripts/components/unified-editor/WorkflowCard/components/StepCard.tsx
@@ -139,7 +139,7 @@ const StepCard = ({
   const isButton = !!onSelectStep;
   const isPlaceholder = sortable.isDragging;
   const isUpgradable =
-    onUpgradeStep &&
+    (onUpgradeStep || onUpgradeStepInStepBundle) &&
     VersionUtils.hasVersionUpgrade(step?.resolvedInfo?.normalizedVersion, step?.resolvedInfo?.versions);
 
   const handleClick = isButton
@@ -181,13 +181,17 @@ const StepCard = ({
   }, [isDragging, isPlaceholder, isButton, isHighlighted]);
 
   const buttonGroup = useMemo(() => {
-    if (!(workflowId || stepBundleId) || isDragging || (!isUpgradable && !onCloneStep && !onDeleteStep)) {
+    if (
+      !(workflowId || stepBundleId) ||
+      isDragging ||
+      (!isUpgradable && !onCloneStep && !onCloneStepInStepBundle && !onDeleteStep && !onDeleteStepInStepBundle)
+    ) {
       return null;
     }
 
     return (
       <ButtonGroup spacing="0" display="none" _groupHover={{ display: 'flex' }}>
-        {isUpgradable && onUpgradeStepInStepBundle && (
+        {isUpgradable && (
           <ControlButton
             size="xs"
             display="none"
@@ -198,16 +202,16 @@ const StepCard = ({
             _groupHover={{ display: 'inline-flex' }}
             onClick={(e) => {
               e.stopPropagation();
-              if (workflowId) {
+              if (workflowId && onUpgradeStep) {
                 onUpgradeStep(workflowId, stepIndex, latestMajor);
               }
-              if (stepBundleId) {
+              if (stepBundleId && onUpgradeStepInStepBundle) {
                 onUpgradeStepInStepBundle(stepBundleId, stepIndex, latestMajor);
               }
             }}
           />
         )}
-        {onCloneStep && onCloneStepInStepBundle && (
+        {(onCloneStep || onCloneStepInStepBundle) && (
           <ControlButton
             size="xs"
             display="none"
@@ -217,16 +221,16 @@ const StepCard = ({
             _groupHover={{ display: 'inline-flex' }}
             onClick={(e) => {
               e.stopPropagation();
-              if (workflowId) {
+              if (workflowId && onCloneStep) {
                 onCloneStep(workflowId, stepIndex);
               }
-              if (stepBundleId) {
+              if (stepBundleId && onCloneStepInStepBundle) {
                 onCloneStepInStepBundle(stepBundleId, stepIndex);
               }
             }}
           />
         )}
-        {onDeleteStep && onDeleteStepInStepBundle && (
+        {(onDeleteStep || onDeleteStepInStepBundle) && (
           <ControlButton
             isDanger
             size="xs"
@@ -237,10 +241,10 @@ const StepCard = ({
             _groupHover={{ display: 'inline-flex' }}
             onClick={(e) => {
               e.stopPropagation();
-              if (workflowId) {
+              if (workflowId && onDeleteStep) {
                 onDeleteStep(workflowId, stepIndex);
               }
-              if (stepBundleId) {
+              if (stepBundleId && onDeleteStepInStepBundle) {
                 onDeleteStepInStepBundle(stepBundleId, stepIndex);
               }
             }}
